// apps/web/src/api/schemas/task.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Farm Recorder API
 * API for GAP-certified farm activity recording and EPCIS traceability
 * OpenAPI spec version: 1.0.0
 */
import type { Uuid } from './uuid';
import type { TaskStatus } from './taskStatus';
import type { Instant } from './instant';

export interface Task {
  id?: Uuid;
  farmId?: Uuid;
  assignedTo?: string;
  title?: string;
  description?: string;
  status?: TaskStatus;
  dueDate?: Instant;
}